Change repository name to Glueby
