+++ Gemfile
@@ -1,6 +1,6 @@
 source "https://rubygems.org"
 
-# Specify your gem's dependencies in tapyrus-contract.gemspec
+# Specify your gem's dependencies in glueby.gemspec
 gemspec
 
 gem "rake", "~> 12.0"
