+++ bin/console
@@ -1,7 +1,7 @@
 #!/usr/bin/env ruby
 
 require "bundler/setup"
-require "tapyrus/contract"
+require "glueby"
 require "tapyrus"
 
 # You can add fixtures and/or initialization code here to make experimenting
