+++ lib/generators/glueby/templates/initializer.rb.erb
@@ -1,0 +1,4 @@
+require 'tapyrus'
+# Edit configuration for connection to tapyrus core
+config = {schema: 'http', host: '127.0.0.1', port: 12381, user: 'user', password: 'pass'}
+Glueby::Contract::RPC.configure(config)
