+++ lib/generators/glueby/templates/timestamp_table.rb.erb
