+++ lib/glueby.rb
@@ -1,0 +1,5 @@
+require "glueby/version"
+
+module Glueby
+  autoload :Contract, 'glueby/contract'
+end
