+++ lib/glueby/contract/active_record.rb
@@ -1,0 +1,9 @@
+require 'active_record'
+
+module Glueby
+  module Contract
+    module AR
+      autoload :Timestamp, 'glueby/contract/active_record/timestamp'
+    end
+  end
+end
