+++ lib/glueby/contract/errors.rb
@@ -1,0 +1,8 @@
+module Glueby
+  module Contract
+    module Errors
+      class InsufficientFunds < StandardError; end
+      class TxAlreadyBroadcasted < StandardError; end
+    end
+  end
+end
