+++ lib/glueby/contract/rpc.rb
@@ -1,0 +1,15 @@
+module Glueby
+  module Contract
+    module RPC
+      module_function
+
+      def client
+        @rpc
+      end
+
+      def configure(config)
+        @rpc = Tapyrus::RPC::TapyrusCoreClient.new(config)
+      end
+    end
+  end
+end
