+++ lib/glueby/version.rb
@@ -1,0 +1,3 @@
+module Glueby
+  VERSION = "0.1.0"
+end
