--- lib/tapyrus/contract/active_record.rb
@@ -1,9 +1,0 @@
-require 'active_record'
-
-module Tapyrus
-  module Contract
-    module AR
-      autoload :Timestamp, 'tapyrus/contract/active_record/timestamp'
-    end
-  end
-end
