--- lib/tapyrus/contract/errors.rb
@@ -1,8 +1,0 @@
-module Tapyrus
-  module Contract
-    module Errors
-      class InsufficientFunds < StandardError; end
-      class TxAlreadyBroadcasted < StandardError; end
-    end
-  end
-end
