--- lib/tapyrus/contract/rpc.rb
@@ -1,15 +1,0 @@
-module Tapyrus
-  module Contract
-    module RPC
-      module_function
-
-      def client
-        @rpc
-      end
-
-      def configure(config)
-        @rpc = Tapyrus::RPC::TapyrusCoreClient.new(config)
-      end
-    end
-  end
-end
