--- lib/tapyrus/contract/version.rb
@@ -1,5 +1,0 @@
-module Tapyrus
-  module Contract
-    VERSION = "0.1.0"
-  end
-end
