--- lib/tasks/tapyrus/contract.rake
@@ -1,9 +1,0 @@
-
-namespace :tapyrus do
-  namespace :contract do
-    desc 'initialize tapyrus contract config'
-    task :install, [] => [:environment] do |_, _|
-      Rails::Generators.invoke("tapyrus:contract:initializer")
-    end
-  end
-end
