+++ spec/glueby/contract_spec.rb
@@ -1,0 +1,5 @@
+RSpec.describe Glueby::Contract do
+  it "has a version number" do
+    expect(Glueby::VERSION).not_to be nil
+  end
+end
