--- spec/tapyrus/contract_spec.rb
@@ -1,5 +1,0 @@
-RSpec.describe Tapyrus::Contract do
-  it "has a version number" do
-    expect(Tapyrus::Contract::VERSION).not_to be nil
-  end
-end
